Create AbstractScheduledJobParameter so plugins can have less boilerplate
